Node: allow Rust `async` blocks to wait on JavaScript futures
Neon provides a way to expose *synchronous* JavaScript functions from
Rust. This means that if Rust wants to wait for the result of a
JavaScript promise, it can at best return a callback to continue its
work when the promise resolves. This does not naturally compose with
Rust's `async`, which works in terms of Futures.

This commit adds a new crate, signal-neon-futures, that provides
functionality for (1) wrapping JavaScript futures so they can be
awaited on in Rust, and (2) producing a JavaScript promise that wraps
a Rust future. It does so by synchronously resuming execution of the
Rust future whenever an awaited JavaScript promise is resolved.

This approach does have some caveats:

- The Neon JavaScript context is still lifetime-scoped, so data must
  be explicitly persisted across await points using
  JsAsyncContext::register_context_data.

- Because execution is synchronous, no other kinds of blocking futures
  are supported. Any await must ultimately bottom out in a JsFuture.